Add new User show route using TDD.
Started by adding the test, which fails indicating a "show" action is
not found in the controller.

After adding the code to the controller, it would still fail indicating
there was no template to show.

So finally added the template which made the test pass.

$ bin/rails test:controllers

We're still missing links to the User show page.

But I got to test entering the `/users/1` and `/users/2` paths manually
into the URL to confirm that was working as expected.