Synchronize comments_count column with actual number of comments.
Generated a skeleton for migration with:
$ bin/rails generate migration PopulatePostCommentsCount

Then populated it with the code to update the comments_count fields.
(Note: I used the `up` method rather than the `change` method originally
populated by `rails generate`.)

Then ran the migration:
$ bin/rails db:migrate

I had to fiddle with it a bit, so I could query migration status with:
$ bin/rails db:migrate:status

And rollback the migration with:
$ bin/rails db:rollback 20200731210404

So that I could apply it again after fixing the code.

Also had to restart the server, in order to flush the in-memory cache.

After finishing this step, I'll remove the tmp/caching-dev.txt file,
since I'm done with testing caching setup at this point.